requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
da-vinci==0.3.0
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-crispy-forms==1.11.2
django-ranged-response==0.2.0
django-recaptcha==2.0.6
django-simple-captcha==0.5.14
django-sitemaps==1.1.1
django-thumbnails==0.4.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
lxml==4.6.3
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==8.2.0
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.6
python-social-auth==0.3.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
shortuuid==1.0.1
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
urllib3==1.25.8
webencodings==0.5.1
yapf==0.31.0
